Added Community Detection implementation

+++ src/de/bigdatapraktikum/twitternews/processing/ChineseWhisperInitialClassMapper.java
@@ -2,13 +2,13 @@ package de.bigdatapraktikum.twitternews.processing;
 
 import org.apache.flink.api.common.functions.MapFunction;
 
-public class ChineseWhisperInitialClassMapper implements MapFunction<String, Integer> {
+public class ChineseWhisperInitialClassMapper implements MapFunction<String, Long> {
 	private static final long serialVersionUID = 1L;
 
-	public static int classId = 1;
+	public static long classId = 1;
 
 	@Override
-	public Integer map(String value) throws Exception {
+	public Long map(String value) throws Exception {
 		return classId++;
 	}
 }
